requirements.txt
Pillow
einops
opencv_python
streamlit
torch
torchvision
transformers
